Changes nonce to long to allow more attempts. No multi-thread support planned.

+++ blockchain/src/main/java/pt/um/lei/masb/blockchain/BlockHeader.java
@@ -53,7 +53,7 @@ public final class BlockHeader implements Sizeable {
     private Instant timeStamp;
 
     @Basic(optional = false)
-    private int nonce;
+    private long nonce;
 
     /**
      * Origin block specialty constructor
